pyproject.toml
[tool.poetry]
name = "robotframework-slack"
version = "0.0.1-alpha"
description = ""
authors = ["vinitrindade88 <vinicius.trindade@diagroup.com>"]
readme = "README.md"
packages = [{include = "robotframework-slack"}]

[tool.poetry.dependencies]
python = "^3.10"
slack-sdk = "^3.21.2"
robotframework = "^6.0.2"
slackclient = "^2.9.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
